+++ package.json
@@ -40,7 +40,7 @@
     "url": "https://github.com/AlreadyBored/node-nodejs-basics/issues"
   },
   "homepage": "https://github.com/AlreadyBored/node-nodejs-basics#readme",
-  "volta": {
+  "volta": { 
     "node": "22.9.0"
   }
 }
