+++ src/constants.js
@@ -1,1 +1,2 @@
 export const FILES_FOLDER = 'src/fs/files';
+export const MODULES_FOLDER = 'src/modules';
