+++ src/modules/esm.mjs
@@ -1,0 +1,41 @@
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as  createServerHttp } from 'http';
+import { promises as fs} from 'fs';
+import './files/c.js';
+import { MODULES_FOLDER } from '../constants.js';
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = await fs.readFile(path.resolve(MODULES_FOLDER,'files/a.json'), 'utf-8');
+} else {
+    unknownObject = await fs.readFile(path.resolve(MODULES_FOLDER,'files/b.json'), 'utf-8');
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${import.meta.url}`);
+console.log(`Path to current directory is ${path.dirname(import.meta.url)}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export default {
+    unknownObject,
+    myServer,
+};
+
